fix(#2755 R4-item6): extract checkout_helpers.rs — checkout.rs back under LOC ceiling
CI failure on 4624646e (Coverage + all Check jobs): tests/file_size_invariant.rs
mcp_handler_files_under_max_loc failed — R3 grew checkout.rs 583→772 LOC, over the
MAX_LOC=750 MCP-handler ceiling (checkout.rs is not grandfathered). My R3 nextest
filter missed this workspace `--tests` invariant.

Extract the self-contained R3 helpers — rollback_response (RollbackOutcome → response
mapping) + sync_marker_contents (marker content-fsync) + the FAIL_MARKER_SYNC cfg(test)
seam — into a new sibling module src/mcp/handlers/ci/checkout_helpers.rs (the same
LOC-relief split pattern as source_resolve.rs). checkout.rs imports them (call sites
unchanged) and is now 704 LOC; checkout_helpers.rs is 77. Tests reference
super::checkout_helpers::* for the moved items. Behaviour byte-identical.

Verified: mcp_handler_files_under_max_loc passes; 41 checkout tests pass.


Agend-Agent: archfix-opus-3
Agend-Task: t-20260712233218081731-15764-2
Agend-Branch: fix/repo-checkout-transactional-submodules
Agend-Issued-At: 2026-07-13T06:49:34.424787+00:00
